+++ lending/src/lib/config.ts
@@ -52,7 +52,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.7.7",
-  genVersion: "2.467.4",
-  userAgent: "speakeasy-sdk/typescript 0.7.7 2.467.4 1.0.0 ryan-lending",
+  sdkVersion: "0.7.8",
+  genVersion: "2.470.1",
+  userAgent: "speakeasy-sdk/typescript 0.7.8 2.470.1 1.0.0 ryan-lending",
 } as const;