+++ lending/RELEASES.md
@@ -799,3 +799,13 @@ Based on:
 - [typescript v0.7.7] lending
 ### Releases
 - [NPM v0.7.7] https://www.npmjs.com/package/ryan-lending/v/0.7.7 - lending
+
+## 2024-12-04 01:04:31
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.451.1 (2.470.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.7.8] lending
+### Releases
+- [NPM v0.7.8] https://www.npmjs.com/package/ryan-lending/v/0.7.8 - lending